[SPARK-43265] Move Error framework to a common utils module
### What changes were proposed in this pull request?

Move Error framework to a common utils module so that we can share it between Spark and Spark Connect without introducing heavy dependencies on Spark Connect module.

### Why are the changes needed?

Reduce Dependencies on Spark Connect.

### Does this PR introduce _any_ user-facing change?

Error framework is internally API so this should be fine.

### How was this patch tested?

Existing UT.

Closes #40931 from amaliujia/SPARK-43265.

Authored-by: Rui Wang <rui.wang@databricks.com>
